cleaning up / adding to gamedata
